rename gem to spbtv_redis_failover, update zk and fix rspec dependencies
SPBTV-205 issue required me to add new gem, google-api-client, to components/subscriptions
which caused a dependency resolution conflict. This gem depended on an old version of "zk",
which in turn depends on an old version of logging gem. I've updated zk to newer version to
resolve this conflict

Also I've fixed some development dependencies for tests to pass